Only run truncate on dev or testing envs

+++ database/seeds/00-cleanup.js
@@ -1,5 +1,10 @@
 exports.seed = async function(knex) {
-  await knex("votes").truncate();
-  await knex("issues").truncate();
-  await knex("users").truncate();
+  if (
+    process.env.DB_ENV === "development" ||
+    process.env.DB_ENV === "testing"
+  ) {
+    await knex("votes").truncate();
+    await knex("issues").truncate();
+    await knex("users").truncate();
+  }
 };
